evals: split into one-file-per-eval
makes it a bit easier to see the pattern
